TO-DO-LIST
A "to-do list" project is a software or application project, usually designed for the purpose of personal or team-based task management. The project I am doing here is a simpler adapted version, offering a tool that helps the user to organize, plan and track daily tasks, projects or jobs. This project uses HTML, CSS, jQuery and Javascript.